# internal/modules/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 九小二 Backend API
  version: 1.0.0-cp1
  description: 九小二产品一期 Go 后端唯一公开契约，覆盖交易、履约、打印、通知、账号运营与酒类合规。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      summary: 分类列表
      security: []
      responses:
        "200":
          description: 分类列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryListEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
  /products:
    get:
      summary: 公共商品列表
      security: []
      parameters:
        - $ref: "#/components/parameters/LocationContext"
        - $ref: "#/components/parameters/AnonymousSession"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - { name: shop_id, in: query, schema: { type: string } }
        - { name: category_id, in: query, schema: { type: string } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: city_code, in: query, schema: { type: string, pattern: "^[0-9]{6}$" } }
        - { name: lat, in: query, schema: { type: number, format: double } }
        - { name: lng, in: query, schema: { type: number, format: double } }
      responses:
        "200":
          description: 公共商品分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductListEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
  /products/{id}:
    get:
      summary: 公共商品详情
      x-auth-mode: public_optional_identity
      x-object-scope: public_optional_location_context
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/LocationContext"
        - $ref: "#/components/parameters/AnonymousSession"
        - { name: shop_id, in: query, schema: { type: string } }
        - { name: city_code, in: query, schema: { type: string, pattern: "^[0-9]{6}$" } }
        - { name: lat, in: query, schema: { type: number, format: double } }
        - { name: lng, in: query, schema: { type: number, format: double } }
      responses:
        "200":
          description: 按服务门店上下文返回的商品详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductDetailEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
...
  /orders:
    get:
      summary: C 端本人订单列表
      x-permission: order:list
      x-object-scope: customer.self_orders
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - { name: status, in: query, schema: { type: string, maxLength: 32 } }
        - { name: order_by, in: query, schema: { type: string, enum: ["created_at desc,id desc"] } }
      responses:
        "200":
          description: 本人订单分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerOrderPageEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      summary: 创建订单并预占库存
      x-permission: order:create
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCreateReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}:
    get:
      summary: 订单详情
      x-permission: order:view
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}/cancel:
    post:
      summary: 取消订单
      x-permission: order:cancel
      x-object-scope: customer.self_order_or_authorized_admin
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}/payments:
    post:
      summary: 创建或复用本人订单的微信支付单
      x-permission: payment:create
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentCreateReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}/payment:
    get:
      summary: 查询本人订单支付事实
      x-permission: payment:view
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}/payment/confirm:
    post:
      summary: 主动向微信确认本人订单支付结果
      description: 小程序支付结果以该接口或服务端回调为准；仅 creating、pending 会请求微信，已成功支付直接返回本地结果。
      x-permission: payment:view
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /payments/{provider}/callbacks:
    post:
      summary: 支付供应商签名回调
      security: []
      parameters:
        - $ref: "#/components/parameters/Provider"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: 供应商 ACK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentCallbackAck" }
        default: { $ref: "#/components/responses/Problem" }
  /orders/{id}/pay/mock:
    post:
      summary: 模拟支付
      x-permission: payment:create
      x-object-scope: customer.self_order
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/MockPay" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /store/orders:
    get:
      summary: 商户授权门店订单
      x-permission: store_order:list
      x-object-scope: merchant_and_authorized_shop
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - { name: shop_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, maxLength: 32 } }
        - { name: keyword, in: query, schema: { type: string, maxLength: 64 } }
        - { name: order_no, in: query, schema: { type: string, maxLength: 64 } }
        - { name: paid_from, in: query, schema: { type: string, format: date-time } }
        - { name: paid_to, in: query, schema: { type: string, format: date-time } }
        - { name: order_by, in: query, schema: { type: string, enum: ["created_at desc,id desc"], default: "created_at desc,id desc" } }
      responses:
        "200":
          description: 授权门店订单分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreOrderPageEnvelope" }
        default: { $ref: "#/components/responses/Problem" }
  /store/orders/{id}:
    get:
      summary: 查看授权门店的商家订单详情
      description: 需要 store_order:view 权限。对非当前商户授权门店的订单统一返回 404，避免暴露订单是否存在；响应不包含完整联系电话、精确坐标和内部支付流水。
      x-permission: store_order:view
      x-object-scope: merchant_and_authorized_shop
      x-sensitive-response: true
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: 商家订单详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreOrderDetailEnvelope" }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
  /admin/orders:
    get:
      summary: 管理后台订单列表
      responses:
        "200": { $ref: "#/components/responses/Page" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/orders/{id}:
    get:
      summary: 管理后台订单详情
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
...
  /admin/orders/{id}/cancel:
    post:
      summary: 管理员受控取消订单
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/OpsReasonVersionReq" } } } }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: application_access JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
    PageToken:
      name: page_token
      in: query
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    LocationContext:
      name: X-Location-Context
      in: header
      schema: { type: string, pattern: "^loc_[A-Za-z0-9_-]{32}$" }
    AnonymousSession:
      name: X-Session-ID
      in: header
      schema: { type: string, minLength: 16, maxLength: 256 }
    Provider:
      name: provider
      in: path
      required: true
      schema: { type: string, enum: [wechat] }
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Page:
      description: 分页请求成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PageEnvelope" }
...
    Problem:
      description: 标准业务或系统错误
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  requestBodies:
...
    MockPay:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/MockPayReq" } } }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { type: object, additionalProperties: true, nullable: true }
        request_id: { type: string }
...
    PageEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          required: [items, next_page_token]
          properties:
            items:
              type: array
              items: { type: object, additionalProperties: true }
            next_page_token: { type: string }
        request_id: { type: string }
...
    OrderCreateReq:
      type: object
      required: [shop_id, address_id, items]
      properties:
        shop_id: { type: string }
        address_id: { type: string }
        items:
          type: array
          items:
            type: object
            required: [shop_product_id, quantity]
            properties:
              shop_product_id: { type: string }
              quantity: { type: integer }
...
    OrderSummary:
      type: object
      additionalProperties: false
      required: [id, order_no, shop_id, status, pay_status, delivery_status, payable_amount, shop_summary, item_summary, item_kind_count, total_quantity, created_at, updated_at]
      properties:
        id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        order_no: { type: string }
        shop_id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        status: { type: string }
        pay_status: { type: string }
        delivery_status: { type: string }
        payable_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        shop_summary: { $ref: "#/components/schemas/ShopSummary" }
        item_summary: { $ref: "#/components/schemas/OrderItemSummary" }
        item_kind_count: { type: integer, minimum: 0 }
        total_quantity: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CustomerOrderPageEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data:
          type: object
          additionalProperties: false
          required: [items, next_page_token]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/OrderSummary" }
            next_page_token: { type: string }
        request_id: { type: string }
...
    Category:
      type: object
      additionalProperties: false
      required: [id, name, sort_order, status, age_restricted]
      properties:
        id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        parent_id: { type: string, pattern: "^[1-9][0-9]{0,19}$" }
        name: { type: string }
        sort_order: { type: integer }
        status: { type: string, enum: [active] }
        age_restricted: { type: boolean }
    CategoryListEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Category" }
        request_id: { type: string }
    ProductListEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data:
          type: object
          additionalProperties: false
          required: [items, next_page_token]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ProductDetail" }
            next_page_token: { type: string }
        request_id: { type: string }
...
    ProductDetailServiceShop:
      type: object
      additionalProperties: false
      required: [id, category_id, context_type, shop_id, shop_product_id, name, sale_price_amount, status, available_qty, purchasable, unavailable_reason, age_restricted]
      properties:
        id: { type: string }
        category_id: { type: string }
        context_type: { type: string, enum: [service_shop] }
        shop_id: { type: string }
        shop_product_id: { type: string }
        name: { type: string }
        brand_name: { type: string }
        spec: { type: string }
        image_url: { type: string }
        description: { type: string }
        sale_price_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        original_price_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        status: { type: string }
        available_qty: { type: integer, minimum: 0 }
        purchasable: { type: boolean }
        unavailable_reason:
          type: string
          nullable: true
          enum: [shop_closed, not_on_sale, out_of_stock]
        delivery_promise: { $ref: "#/components/schemas/ProductDeliveryPromise" }
        age_restricted: { type: boolean }
...
    ProductDetail:
      oneOf:
        - $ref: "#/components/schemas/ProductDetailServiceShop"
        - $ref: "#/components/schemas/ProductDetailNoServiceShop"
      discriminator:
        propertyName: context_type
        mapping:
          service_shop: "#/components/schemas/ProductDetailServiceShop"
          no_service_shop: "#/components/schemas/ProductDetailNoServiceShop"
    ProductDetailEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data: { $ref: "#/components/schemas/ProductDetail" }
        request_id: { type: string }
...
    MockPayReq:
      type: object
      required: [channel]
      properties:
        channel: { type: string, enum: [mock] }
    PaymentCreateReq:
      type: object
      required: [provider, client_type]
      properties:
        provider: { type: string, enum: [wechat] }
        client_type: { type: string, enum: [miniapp] }
        return_context: { type: object, additionalProperties: true }
...
    PaymentCallbackAck:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [SUCCESS, FAIL] }
        message: { type: string }
...
    StoreOrderSummary:
      type: object
      additionalProperties: false
      description: 商家订单列表的最小核单投影；不包含顾客 ID、完整联系方式、完整地址或全部商品。
      required: [id, order_no, shop_id, status, pay_status, delivery_status, payable_amount, shop_summary, item_summary, item_kind_count, total_quantity, address_summary, customer_contact_mask, has_remark, version, created_at, updated_at]
      properties:
        id: { type: string }
        order_no: { type: string }
        shop_id: { type: string }
        status: { type: string }
        pay_status: { type: string }
        delivery_status: { type: string }
        payable_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        shop_summary: { $ref: "#/components/schemas/StoreShopSummary" }
        item_summary: { $ref: "#/components/schemas/StoreOrderItemSummary" }
        item_kind_count: { type: integer, minimum: 0 }
        total_quantity: { type: integer, minimum: 0 }
        address_summary: { type: string, maxLength: 512, x-sensitive: pii_masked }
        customer_contact_mask: { type: string, x-sensitive: pii_masked }
        has_remark: { type: boolean }
        version: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        paid_at: { type: string, format: date-time }
    StoreOrderPageEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data:
          type: object
          additionalProperties: false
          required: [items, next_page_token]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/StoreOrderSummary" }
            next_page_token: { type: string }
        request_id: { type: string }
...
    StoreOrderDetail:
      type: object
      additionalProperties: false
      required: [id, order_no, shop_id, status, pay_status, delivery_status, payable_amount, shop_summary, item_summary, item_kind_count, total_quantity, created_at, items, address_snapshot, remark, delivery_promise, compliance_summary, goods_amount, discount_amount, delivery_fee_amount, paid_amount, version, payment_summary, customer_contact_mask, delivery_summary, recent_logs]
      properties:
        id: { type: string }
        order_no: { type: string }
        shop_id: { type: string }
        status: { type: string }
        pay_status: { type: string }
        delivery_status: { type: string }
        payable_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        shop_summary: { $ref: "#/components/schemas/StoreShopSummary" }
        item_summary: { $ref: "#/components/schemas/StoreOrderItemSummary" }
        item_kind_count: { type: integer, minimum: 0 }
        total_quantity: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/StoreOrderItem" }
        address_snapshot: { $ref: "#/components/schemas/StoreOrderAddressSnapshot" }
        remark: { type: string }
        delivery_promise:
          allOf:
            - $ref: "#/components/schemas/StoreDeliveryPromise"
          nullable: true
        compliance_summary: { $ref: "#/components/schemas/StoreComplianceSummary" }
        goods_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        discount_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        delivery_fee_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        paid_amount: { type: integer, format: int64, minimum: 0, description: 人民币分 }
        cancel_source: { type: string }
        cancel_reason_code: { type: string }
        version: { type: integer, minimum: 0 }
        expires_at: { type: string, format: date-time }
        paid_at: { type: string, format: date-time }
        cancelled_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        payment_summary:
          allOf:
            - $ref: "#/components/schemas/StorePaymentSummary"
          nullable: true
        customer_contact_mask: { type: string, description: 已脱敏顾客联系方式, x-sensitive: pii_masked }
        delivery_summary:
          allOf:
            - $ref: "#/components/schemas/StoreDeliverySummary"
          nullable: true
        recent_logs:
          type: array
          items: { $ref: "#/components/schemas/StoreOrderLogSummary" }
    StoreOrderDetailEnvelope:
      type: object
      additionalProperties: false
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, enum: [0] }
        message: { type: string, example: ok }
        data: { $ref: "#/components/schemas/StoreOrderDetail" }
        request_id: { type: string }
...
    OpsReasonVersionReq:
      type: object
      required: [reason_code, reason, expected_version]
      properties:
        reason_code: { type: string, maxLength: 64 }
        reason: { type: string, minLength: 2, maxLength: 255 }
        expected_version: { type: integer, minimum: 1 }
...
    ProblemDetails:
      type: object
      required: [type, title, status, error_code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error_code:
          type: string
          description: 幂等写请求同 key 异参固定返回 IDEMPOTENCY_KEY_REUSED；同参请求仍处于有效处理租约时固定返回 IDEMPOTENCY_IN_PROGRESS。
        request_id: { type: string }